==> notebooks/2_get_journals.ipynb <==
import os, io, json, numpy as np, codecs
import pandas as pd
import pickle
import urllib.request


this_month='1711'
data_path='../data/'+this_month

    url="http://api.crossref.org/journals?rows=1"
    response = urllib.request.urlopen(url)
    html = response.read()    
    json_result=json.loads(html)
    count=json_result['message']['total-results']

dic=[]
for it in range(int(count/1000)+1): # change
    print (it, end=','),
    offset = it * 1000
    p={'rows' : '1000', 'offset' :str(offset),'mailto':'mahdi.moqri@gmail.com'}
    parameters=urllib.parse.urlencode(p)
    url='http://api.crossref.org/journals?'+parameters
    response = urllib.request.urlopen(url).read()
    json_result=json.loads(response)
    items=json_result['message']['items']
    for item in items:
        if len(item['ISSN'])>0 and 'title' in item and 'publisher' in item: 
            journal={}
            journal['e:title']=item['title'].strip()
            journal['e:publisher']=item['publisher'] 
            journal['e:issn']=item['ISSN']
            journal['e:subjects']=[]
            for subject in item['subjects']:
                journal['e:subjects'].append(subject['name'].strip())
            if journal not in dic:
                dic.append(journal)
dic_c=dic      

count

with open('../store/wd/journals.tab','w') as f:
    for j in dic:
        f.write (','.join(j['e:issn'])+'\t'+j['e:title']+'\t'+j['e:publisher']+'\n')

with open ('../store/wd/wd_issns.csv') as f:
    wd_issns=f.readlines()
    wd_issns=[issn.strip('\n') for issn in wd_issns]
with open('../store/wd/missing.tab','w') as f:
    for j in dic:
        if j['e:issn'][0] in wd_issns:
            continue
        if len(j['e:issn'])==2 and j['e:issn'][1] in wd_issns:
            continue
        f.write (','.join(j['e:issn'])+'\t'+j['e:title']+'\t'+j['e:publisher']+'\n')

with open(data_path +'/journals.pkl', 'wb') as f:
    pickle.dump(dic, f, pickle.HIGHEST_PROTOCOL)

subs=[]
for j in dic:
    for sub in j['e:subjects']:
        if sub not in subs:
            subs.append(sub)

len(subs)
