==> notebooks/6_top20.ipynb <==
%matplotlib inline
import os, io, json, urllib, numpy as np, codecs
import glob, pandas as pd
pd.options.display.max_rows = 999

this_month='1711'
data_path='../data/'+this_month

df=pd.read_csv(data_path+'/cleaned.csv')
df.shape

df=df[~((df.year==2014) & (df.mm<int(this_month[2:4])))]
df.shape

df=df[(df['year']<=2017) & (df['year']>=2014)]
df.shape

unique_subjects=pd.read_csv('../store/unique_subjects.csv',header=None)
unique_subjects=unique_subjects[0].values

with open('../store/issn_subject.json') as data_file:    
    issn_subject = json.load(data_file)

df2=df[['doi','main_issn','citation']]

join_subjects=df2['main_issn'].apply(lambda x: ';'.join(issn_subject[x]) if x in issn_subject else '')

df2['subjects']=join_subjects.values

df2['main_issn'].apply(lambda x: ';'.join(issn_subject[x]) if x in issn_subject else '').shape

unique_subjects=[]
all_subjects=df2['subjects'].values
for journal_subjects in all_subjects:
    subjects=journal_subjects.split(';')
    for subject in subjects:
        if subject not in unique_subjects: unique_subjects.append(subject)

len(unique_subjects)

top_articles=[]
counter=0
for subject in unique_subjects:
    if subject=='':
        df_subject=df2.sort_values('citation', ascending= False)[:20]
    else:
        df_subject=df2[df2.subjects.apply(lambda x: subject in x.split(';'))].sort_values('citation', ascending= False)[:20]
    top_dois=df_subject[['doi','main_issn','citation']].to_dict('list')
    zips=zip(top_dois['doi'],top_dois['main_issn'],top_dois['citation'])
    items=[];
    for doi,issn,citation in zips:
        item={}; item_cleaned={}
        url='http://api.crossref.org/works/'+doi
        while True:
            try:
                response = urllib.urlopen(url).read()    
            except Exception,e: 
                print str(e)
                time.sleep(1)
                continue
            break
        json_result=json.loads(response)
        item = json_result['message']
        year_print=[9999,99,99];year_online=[9999,99,99];
        if 'title' in item:
            title=item['title'][0]
        if 'published-print' in item:
            year_print=item['published-print']['date-parts'][0]                
        if 'published-online' in item:
            year_online=item['published-online']['date-parts'][0]
        year=min(year_print,year_online)
        if 'DOI' in item: doi=item['DOI']
        if 'container-title' in item: journal=','.join(item['container-title'])
        if 'publisher' in item: pub=item['publisher']; pub=''if pub==None else pub
        if 'references-count' in item: ref=str(item['references-count'])
        authors=[]
        if 'author' in item: 
            for author in item['author']:
                if 'given' in author:
                    author_name=author['given']
                if 'family' in author:
                    author_name+=' '+author['family']
                authors.append(author_name)
        if len(authors)>5: authors=authors[:5]
        if len(year)>1 :
            month=str(year[1]) 
        else: month=''
        item_cleaned={'@id':doi,
                      'e:title':title,
                      'e:year':str(year[0])+'-'+month,
                      'e:cite':np.asscalar(citation),
                      'e:issn':issn, 
                      'e:journal':journal,
                      'e:pub':pub,
                      'e:authors':authors,
                      'e:subject':subject,
                      'e:month':this_month} 
        if item_cleaned not in top_articles:
            top_articles.append(item_cleaned)
    print counter,
    counter+=1            

with open('..\web\cats.json') as data_file:    
    cat_subject=json.load(data_file)

sub_cat={}
for cat in cat_subject.keys():
    for sub in cat_subject[cat].split(';'):
        sub_cat[sub]=cat

top_articles_new=[]
for a in top_articles:
    a['e:subject']=a['e:subject'].strip()
    a['e:category']=''
    subject=a['e:subject']
    if subject!='':
        cat=sub_cat[subject]
        a['e:category']=sub_cat[subject]
    top_articles_new.append(a)

with open('../jsonld/'+this_month+'/top_articles.jsonld', 'w') as fp:
    json.dump(top_articles_new, fp)    

for subject in unique_subjects:
    df_lic=df2[df2.subjects.apply(lambda x: subject in x.split(';'))]
    top_lics=df_lic[['lic','count']].groupby('lic').count().sort_values('count',ascending=False).head().reset_index().to_dict('list')
    zips=zip(top_lics['lic'],top_lics['count'])
    total_articles=str(df_lic.shape[0])
    items=[{'count':total_articles,'lic':'all'}];
    for lic,count in zips:
        item={}; 
        item={'lic':lic,'count':str(count)}
        items.append(item)  
    with open('../jsonld/lic_'+subject+'.json', 'w') as fp:
        json.dump(items, fp)
